crawl4ai/requirements.txt
crawl4ai
python-docx
reportlab
openpyxl